Track max seqno of deleted items per vbucket
In order to guarantee monotonically increasing seqno values even for
recreated items i.e., items that are deleted and reintroduced using
an add or a set command, we track the max seqno of all deleted items
seen so far. Then, when an item is reintroduced, we pick this max
seqno value, increment it by 1 and assign it to the "new" item.
Although conservative, this guarantees that if an item is thus
reintroduced, it gets a seqno that's higher than its previous value.

Change-Id: Ibb30fc5a90db21e05bf1043bfcb1c1050f3913ce
Reviewed-on: http://review.couchbase.org/11901
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>