MB-4551 Record the last closed checkpoint upon cursor registration.
When the cursor for the registered TAP client is activated in the
checkpoint datastructure, the last closed checkpoint id is recorded
in the cursor. If the client is only interested in receiving items
from closed checkpoints only, its cursor keeps walking until
reaching to the end of that closed checkpoint, and then its TAP
connection will be closed.

Change-Id: I56c564fbd880fc6e87804d1e606708ed80fb66c6
Reviewed-on: http://review.couchbase.org/11943
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>