MB-4553 Support delete operations during online restore.
For fast recovery, we allow the online restore to be interleaved
with delete as well as set operations. For this, deletions of
existing or non-existing keys, which are received from the upstream
master via TAP or the normal clients, are maintained to skip the
same keys from the restore backup files.

Upon the restore completion, the list of keys deleted will be
removed.

Change-Id: I9a5d73b77fdb0960fc45582745fc08f1b396f5c8
Reviewed-on: http://review.couchbase.org/12001
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>