Add validation for missing checkpoints while merging backup
If we specify a valid existing backup file as output using
-o outputfile.mbb, the current mbbackup-merge-incremental
script does not do any valdation for missing checkpoints.
This patch adds validation for merging input backup files
to an existing valid backup file which is specified as an
output file. It throws an error message and exits when
checkpoints are missing.

Change-Id: I70e442f4aae7f9ce3dfa5487ac31d0e6850780fd
Reviewed-on: http://review.couchbase.org/12004
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>