Remove timing histogram on store commands.
This is ostensibly a useful stat, but in practice it slows down the
command more than it gives us a feel for how fast it is.  The stat was
slightly misleading as it is really the amount of time it takes a
pre-delivered value to be linked into an in-memory hashtable.  This is
something few people actually want to know and I think we'd rather the
application be faster than talk about how fast it is.

Change-Id: I7d15564cebdc0c736344db1f5aae07bf0306e263
Reviewed-on: http://review.couchbase.org/12063
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>