Introduce unlocked_restoreMeta() method in StoredValue
Similar to the unlocked_restoreValue() method, this method will be
called upon completion of a background fetch of a temporary item's
metadata. After restoring the metadata, this method also sets it to
expire after a predetermined amount of time (default is 60 secs). Any
number of requests within this window will be served with the fetched
metadata.

Change-Id: I7153c64bd35a8b57c0e4d4186008c505fdfd357f
Reviewed-on: http://review.couchbase.org/12207
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>