Expire temporary items created in response to get-meta requests
Upon restoring deleted items from the persistent store, we set them
to expire after a predetermined amount of time (default is 60 secs).

This change set modifies the Deleter class to also look for temporary
items and delete them when the following conditions are met:
1. They are either in the meta_available or non_existent states i.e.,
   the background fetches for these items have completed. If a fetch
   is in process, we let it finish.
2. They have expired.

Also, upon deleting temporary items, we don't put them on the check-
point queues as don't need to be either persisted or replicate to
another node.

Change-Id: Ife51b5b37e72f63db817f4c7519f5e4da3cd8eaa
Reviewed-on: http://review.couchbase.org/12211
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>