Don't copy the duplicate meta data into the engine_specific field
The meta data of each item was previously constructed by
allocating 22 bytes of a byte array in Item class and passing its
pointer to the memcached through the engine_specific field.

As all the meta data fields except for an item's seqno are already
included in an TAP message, we only need to pass the seqno to the
memcached via the engine_specific field. This can also reduce the
memory overhead because Item instances are used in memcached,
flusher queue, checkpoints, etc.

Change-Id: Iadb2b34aa9b14a5f0dc506f6a247544283a9f1a2
Reviewed-on: http://review.couchbase.org/12293
Reviewed-by: Srinivas Vadlamani <srinivas@couchbase.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>