Schedule background fetch for non-existent items in getMetaData()
If the key being queried doesn't exist in memory, create a temporary
item and schedule a background fetch for its metadata from the
persistent store. Until the fetch completes, the temporary item will
be in meta_pending state during which further invocations of this
function will be blocked. After the fetch completes and the state has
transitioned to either meta_available or non_existent, the correct
state will be returned.

NOTE: We currently return ewouldblock in the meta_pending state This
isn't optimal from a performance point of view. This will change
shortly -- we will implement an asynchronous mechanism so that the
connection doesn't have to block until the fetch completes.

Change-Id: I9a761b916920ffe00291251873f6c26bcf96dfed
Reviewed-on: http://review.couchbase.org/12456
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>