MB-4619 Set the empty open checkpoint id to the master's one.
If the slave has the empty open checkpoint and receives a new
open checkpoint id because of collapsing checkpoints in the
upstream master, the slave can simply set its open checkpoint
id to the new id from the master. This will prevent checkpoint
id missings in the incremental backup files.

Change-Id: If1d562f699c0dc2f545cff217d01d6fc54232c43
Reviewed-on: http://review.couchbase.org/12358
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>