Only record and item as deleted for its persistence callback
We increment the stat for an item being deleted from a vbucket
when after we pass the callback to the kvstore. This will cause
the count to get screwed up if the delete fails and the delete
has to be retried. This stat should be incremented when the
delete actually succeeds.

Change-Id: I368bcdab1dc5d3839da2f0e8473e7f582b434a30
Reviewed-on: http://review.couchbase.org/12541
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>