Optimize restore query to improve restore performance
Most of the validations and ordering of mutations according to
checkpoint id can be performed outside of ep-engine.
The RestoreManager should be able to simply restore the given
backup file without doing lot of validations. This patch adds an
ep-engine parameter `restore_file_checks=false`, which can be used
to enable or disable the validations on the backup file. The default
value for restore_file_checks is true.

It is seen that, upto 95% of total time for a file restore operation
is consumed for executing the first sqlite3_step() statement due to
the expensive query. Use `restore_file_checks=false` parameter with
the following constraints:
 - The backup file contains at most one checkpoint
 - The restores are performed by doing restore in the reverse
 chronological order by which the backups are taken.

 Usage of restore_file_checks=false will result upto 95% faster
 online restore.

Change-Id: Ia752bd38c39680f848d5a9494307a19e420338f5
Reviewed-on: http://review.couchbase.org/12563
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>