MB-4661 Don't skip dead vbuckets in persisting items into disk.
During the rebalance, a given dead vbucket can be set to replica
state before the janitor issues a vbucket deletion command for
that dead vbucket. If the dead vbucket has the same open
open checkpoint id as the active vbucket, the vbucket reset and
backfill won't be required. Therefore, the persistence tasks on
dead vbuckets should not be skipped. If those dead vbuckets are
deleted from memory by the janitor that runs every 10 sec,
their persistence tasks will be then skipped.

Change-Id: Ia0f164a3cd8c7d6e959914c9f4f01cbf5e9ef022
Reviewed-on: http://review.couchbase.org/12591
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>