MB-4671 Fix to the deadlock between restore process and flusher.
When ep-engine is running in restore-mode and keys are getting
restored, flusher processes incoming delete operations
in PersistenceCallback. The engine's restoreItem method and
PersistenceCallback grabs the hashtable bucket lock and restore
mode lock. The order in which the locks are acquired is wrong
in the engine's restoreItem method, which can cause the
dealock between the restore and flusher threads.

Change-Id: I0d25be31fe86aed8b3cf62dcb00bfc3ac22fc636
Reviewed-on: http://review.couchbase.org/12596
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>