Refactored soft delete functions in stored value
This commit rolls delete and delete with meta into a single function
in order to get rid of duplicate code. Soft delete also now takes
a StoredValue parameter instead of a string. This will make it so
that we only have to do one hash table look up instead of two.

Change-Id: Ifd70c50f231f4a3a74b5a40c414a686f6d645d46
Reviewed-on: http://review.couchbase.org/12599
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>