Return CAS along with enoent in get_meta()
When we receive a get_meta request for a non-existent key, we create
a temporary item and schedule a background fetch to check whether the
item exists as deleted in Couch. After the background fetch completes
we return either the deleted revid or enoent (if the key never wasn't
found in Couch). This change set also returns the CAS initially
assigned to the temporary item. In case of XDCR, CAPI will use the
CAS to issue a subsequent add_/delete_with_meta command.

Change-Id: Ice47710713dc05e4d2cd25fbc9865ae97a9f2328
Reviewed-on: http://review.couchbase.org/13754
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>