MB-4714: Reset connection if mccouch stops responding
If mccouch crashes after we send a request and before we receive
a response then ep-engine will wait forever for a response that
won't come. This change introduces a new parameter called
couch_response_timeout that will cause ep-engine to re-establish
a connection to mccouch if it a request doesn't come back for the
given amount of time specified in couch_response_timeout. By
default this parameter is set to 10 minute.

Change-Id: Iede592f378d3ed73c0f7aed1a2986148cfb25d0b
Reviewed-on: http://review.couchbase.org/12720
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>