MB-4707 Merge all checkpoints if chk id from the master is smaller
The checkpoint id from the new master could be smaller than the
open checkpoint id on the slave right after rebalance because the
new vbucket map generated for rebalance is sometimes inefficient
and promotes one of the replica vbuckets, which does not catch up
its original master, to the new master. In this case, the slave
simply removes checkpoints whose ids are larger than the
checkpoint id from the new master. However, delete operations in
the checkpoints removed won't be persisted.

To resolve the above issue, this change collapses the checkpoints
on the slave into a single checkpoint so that all delete
opeartions can be persisted and does not break any view
consistency.

Change-Id: Ie232d8da3ca217cda35e609d0a8a84bcbfd3321b
Reviewed-on: http://review.couchbase.org/12853
Reviewed-by: Jin Lim <jin@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>