Don't reschedule checkpoint items only for the registered TAP.
If the TAP connection is for the registered client that is only
interested in items from the closed checkpoints, only checkpoint
items should not be rescheduled because its cursor is going to
be reset to the beginning of its current checkpoint upon
the reconnection. On the other hand, other TAP opaque messages
should be rescheduled for the registered client.

Change-Id: I7f0a267b935d7efe7bab5ca776fd111b74b96abb
Reviewed-on: http://review.couchbase.org/13069
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>