MB-4772 Fix to handle backfill for closed checkpoint only tap
rollback() method was clearing the taplog items when a tap
connection is restablished in order to reset the cursor back to
the start of current checkpoint. Due to this condition, when
backfilled items are added to the taplog, it was getting cleared and
hence backilled mutations where never sent through the tapconnection
since it is empty. This patch adds a condition to handle backfill
with closed checkpoint only tap cursor.

Change-Id: I2fdcda9fe3702f8efd2147b799cfbab142babf85
Reviewed-on: http://review.couchbase.org/13092
Reviewed-by: Manik Taneja <mtaneja@zynga.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>