Disable the checkpoint extension mode if the memory usage is high
The cluster manager enables the checkpoint extension mode for
active vbuckets that are affected by the maintenance mode, so that
the full backfill from those active vbuckets can be avoided if
possible when the node in the maintenance mode rejoins the
cluster.

However, if the memory usage is still above the high water mark
and there were few items evicted even after the item pager runs,
the extension mode on active vbuckets should be disabled, so that
their open checkpoints can be closed and their non-dirty items
can be evicted.

Change-Id: Id44b6a91f3f3843bb4716e9ddcc662b06c71819f
Reviewed-on: http://review.couchbase.org/13176
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>