MB-4813 Maintain TAP backfill queue and backlog size separately.
Backfill rate from disk is usually slower than the replication
rate if disk IO becomes bottleneck. In this case, the backfill
backlog stat is usually ranged in a very small number as it is
currently based on the size of backfill queue in the TAP
connection.

To address the above issue, we maintain two separate stats for
backfill. The first stat is the backfill queue size that was
previously used to represent the total backfill backlogs.
The second stat is the total backfill backlogs that are
estimated right before memory/disk backfill task is scheduled,
and then updated while sending backfill items to the destination.

Change-Id: Icf25dd1cebd872a107646c30d65db084286eed0d
Reviewed-on: http://review.couchbase.org/13344
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>