Log compactor that compacts a mutation log periodically.
This is the initial implementation of the mutation log compactor.
The compactor is scheduled periodically and checks the
following conditions to see if the compaction is required for the
current mutation log file:

1) current mutation log size > max log size allowed,
2) num of total logged items > max entry num allowed,
3) write queue size < queue cap allowed,

If all the above conditions are satisfied, the compactor suspends
the persistence and then dumps the entire hash table into a new
log file and finally renames it to the original log file.

Change-Id: Ibfbe3bca339574bebd94de9ac0958005e1193e76
Reviewed-on: http://review.couchbase.org/13507
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>