Don't maintain replica counter field from StoredValue class.
We don't need to maintain the replica counter field in StoredValue
class anymore as sync commmand is not supported anymore.
We're going to provide the new approach for ensuring if an item
is replicated or not.

The memory overhead per StoredValue instance is reduced by 8 bytes
(i.e., 71 bytes to 63 bytes).

Change-Id: Ic3aaa2ac996066ac4a833a29b4c015b4bd2f2e6d
Reviewed-on: http://review.couchbase.org/13585
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>