mb-4857 prevent to free object prematurely
InvalidItemDbPager is allocated in the constructor. And it will be
used right after warmup. And it is freed in the initialize(). If
dispatcher schedules the task ahead of flusher task, the original
object will be freed prematurely before it is referenced again after
warmup. We use share_ptr to keep honest of object pointer references.

Change-Id: Iaf59e113df58a5f5a53e9cf35b6cd38d9a2afc19
[alk@tut.by: whitespace-only M-x align]
Signed-off-by: Aliaksey Kandratsenka <alk@tut.by>
Reviewed-on: http://review.couchbase.org/13817
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Bin Cui <bin.cui@gmail.com>