Help valgrind to not report false positives for uninit reads
Running valgrind on my mac reports multipe uninitialized reads
caused by the mutexes / syncobjects. We're initializing these
objects with the pthread_xxx_init methods, but it seems to leave
some of the internals uninitialized (and probably read out a
word including these bytes). Add a --enable-valgrind so that we
can define VALGRIND and memset the variables in these cases.

Change-Id: Ief294963fe1f03ad18b5b011b27d19860fe50f9d
Reviewed-on: http://review.couchbase.org/13916
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>