MB-4898 Rebuild mutation log with the database if the log is empty
There is a case where the database has some data, but the mutation
log is empty when the warmup starts (e.g., cluster upgrade).
In this case, the mutation log should be reconstructed from
database during warmup. Otherwise, data loss can happen when the
node is restarted, because the mutation log might not contain all
the entries.

Change-Id: I2fdbb9458c54f6e074b15fb18f7d757234bcc3c9
Reviewed-on: http://review.couchbase.org/14007
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>