don't assume ETIMEDOUT is 110
There's a problem in windows headers and/or pthreads for windows with
different values of ETIMEDOUT. Previous solution was a bit naive and
broken. It just defined 110 as CB_ETIMEDOUT and added extra case.

My compiler (quite correctly) have warning there so I was forced to
fix.

New fix hides this sad fact of two ETIMEDOUT values inside inline
wrapper for pthread_cond_timedwait. Without making assumption that 110
is ETIMEDOUT on all platforms.

Change-Id: I8dd456a92960d8247d2757c01c0d5684d4c79d89
Reviewed-on: http://review.couchbase.org/14318
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>