MB-4990 Fix to the bug in resetting a cursor for backfill
If backfill is required for a given vbucket, we remove the
corresponding checkpoint cursor from that vbucket and then set
the cursor to the beginning of that vbucket's open checkpoint when
the backfill is actually started. In this way, we can send live
mutations to the destination by moving the cursor after backfill
is completed.

There is a bug in removing a checkpoint cursor temporarily from
the vbucket. The cursor has to be removed when we are sure that
the backfill is going to be scheduled for a given vbucket.

Change-Id: Ib56754c14cb7d4f742835b477490685021f6f6b6
Reviewed-on: http://review.couchbase.org/14509
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>