MB-4951: Update stats using a single thread
Every time we try to add an item to the database, check to see
if there is room in checkpoint queues, try to see of we have
enough memory for backfill, ... we need to check that we aren't
using more memory than we have allocated to use. This process
required getting stats from tcmalloc and getting stats required
going through a few spin locks and this resulted in a lot of
contention. This fix adds a new thread to the MemoryTracker that
will update the stats every 1/4 second in order to relieve
contention around the spin locks in tcmalloc.

Change-Id: Ibd54bfe1e1cbb6e201d9759d68dbf3399aaf45f0
Reviewed-on: http://review.couchbase.org/14544
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>