MB-4972: Bucket memory fragmentation has better accounting
We encountered an issue that caused a bucket to run out of memory
due to fragmentation calculation when deleteing buckets. This happened
because we accounted for fragmentation as a percentage of total
allocated memory. This becomes an issue if you have a lot of buckets
using all available memory and then delete all but one of the buckets.
If this scenario occurs then all of the fragmenation from the other
buckets can be given to the remaining bucket and render that bucket
useless. What should happen in this case is the fragmentation should
be spread across then entire processes memory. This change adjusts
the algorithm to do exactly this.

Change-Id: I3497b3f03ed332191a76720156348632b1506598
Reviewed-on: http://review.couchbase.org/14569
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-on: http://review.couchbase.org/14916
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>