MB-4972: Bucket memory fragmentation has better accounting
We encountered an issue that caused a bucket to run out of memory
due to fragmentation calculation when deleteing buckets. This happened
because we accounted for fragmentation as a percentage of total
allocated memory. This becomes an issue if you have a lot of buckets
using all available memory and then delete all but one of the buckets.
If this scenario occurs then all of the fragmenation from the other
buckets can be given to the remaining bucket and render that bucket
useless. What should happen in this case is the fragmentation should
be spread across then entire processes memory. This change adjusts
the algorithm to do exactly this.

Change-Id: I8aab73cc7af0d2bae1837137211eb94b57d6841a
Reviewed-on: http://review.couchbase.org/14569
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>