MB-5125 Revert "Remove unnecessary dedup checks from beginFlush()"
This reverts commit 3c8d17ff067f36e148efef96cab4ffb451e93391.

We observed the seg fault crash from std::sort() if we sort
QueuedItem instances by std::string. We even replaced RCPtr with
shared_ptr, but still had the same issue.

Using std::set is more expensive, but we observed that it has
little impact on the frontend performance or write througput.

Change-Id: Iec3a6bbedd119e37c72e54de2de7cf844e1982ce
Reviewed-on: http://review.couchbase.org/15205
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Jin Lim <jin@couchbase.com>