Encode expiration time instead of length in item revision ids
A recent change in couchstore encodes expiration time instead of
length in the revision ids of documents. This change set makes the
corresponding changes to ep-engine so that a get_meta request will
return not the length but the expiration time.

Also, as part of this change, a struct to hold the revision id
components -- namely, the seqno, the cas, the expiration time and the
flags -- is added to simplify encoding and decoding.

Change-Id: Ic24ce9946131f886cc554d17baefdbda2a9946eb
Reviewed-on: http://review.couchbase.org/15285
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>