Do not set expiration time for temporary items
Previously, after the bgfetch for a temporary item completed, we used
to set it to expire after 1 hr. However, with the recent change of
encoding expiration time instead of length in an item's revid, it is
no longer safe to do this.

The new approach is as follows: After the bgfetch for metadata
completes, we set its seqno, cas, expiration time and flags
attributes. These are then used to service requests for get_meta.
Anytime either the item pager or the expiry pager runs, we look for
temporary items and delete them from memory. If CAPI does not finish
conflict resolution during this time, it will retry.

Change-Id: Ie4bc662c39e085c4bde73f279b845de5cd1723a3
Reviewed-on: http://review.couchbase.org/15286
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>
Reviewed-by: Jin Lim <jin@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>