MB-5193 Don't reschedule backfill upon resending an OPAQUE message
A backfill task on the source node is scheduled upon sending
the destination TAP_OPAQUE_INITIAL_VBUCKET_STREAM message that
indicates the beginning of backfill for a given vbucket.
If this opaque message should be resent because it is not acked
from the destination, the source node doesn't need to schedule
the backfill task again, but instead simply resend the opaque
message to the destination.

Change-Id: Id6b25385eefdabf34d0c30bc41851435ad4454e0
Reviewed-on: http://review.couchbase.org/15544
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>