MB-5193 Don't reschedule backfill upon resending an OPAQUE message
A backfill task on the source node is scheduled upon sending
the destination TAP_OPAQUE_INITIAL_VBUCKET_STREAM message that
indicates the beginning of backfill for a given vbucket.
If this opaque message should be resent because it is not acked
from the destination, the source node doesn't need to schedule
the backfill task again, but instead simply resend the opaque
message to the destination.

Change-Id: Idd29728ecff357927064f086d507ae82c3b78c9a
Reviewed-on: http://review.couchbase.org/15544
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-on: http://review.couchbase.org/15571
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>