MB-5227 check vbucket pointer before updating its stat
The change is to fix the segfault that can occur during a callback
of an item deletion. When an item deletion is complete and its callback
is invoked, it is possible that the vbucket of the item has already
been deleted thus removed from the vbucketmap. In such case the callback
should not attempt to update any stat of the deleted vbucket (opsDelete).

Change-Id: Ie189358a26b542e91f85617c8d12015e1d5fbba6
Reviewed-on: http://review.couchbase.org/15662
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Liang Guo <liang@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>