MB-5176: Modify the item eviction policy
This change modifies the item eviction policy to run under
two different conditions. We now evict when the total amount
of allocated items reaches the high watermark or when the
total allocated items + fragmentation is above the maximum
mutation threshold.

Before this change fragmentation could cause the item eviction
process to evict a large percentage of the items in the cache
because doing a deallocation in tcmalloc doesn't neccessarily
mean that memory is freed back to the allocator for general
purpose use. This change greatly reduces the possibility that
this situation can happen and as a result keeps the resident
ratio much high in Couchbase nodes.

Change-Id: I4a7b51d5025272fe69cf83421df733ebab568a20
Reviewed-on: http://review.couchbase.org/15769
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>