MB-4418: tap_throttle_queue_cap in cbflushctl takes 'infinite' argument
The infinite argument will map to -1 for the tap_throttle_queue_cap. We
are adding this mapping because in order for cbflushctl to be able to
parse -1 it would require doing '-- -1' and this is confusing and not
intuitive for ordinary users. Those who are familiar with the -- will
however still be able to use it.

Change-Id: I934f3d2c667cb81f54e7830f13b58368c5167915
Reviewed-on: http://review.couchbase.org/15895
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>