MB-5401 Reopen database upon failures from the database layer
Failures from the underlying database layer can happen by various
reasons. For example, if a database file is deleted by mistake or
a database directory is mounted to the NFS that sometimes has a
connection issue, there would be failures from the underlying
database layer. In this case, we should reopen database to get
a new handle.

Change-Id: Iaf67d7eac195c2e1c4a5c71ca26b6d89b11adfff
Reviewed-on: http://review.couchbase.org/16508
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-on: http://review.couchbase.org/16674
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>