MB-5401 Reopen database upon failures from the database layer
Failures from the underlying database layer can happen by various
reasons. For example, if a database file is deleted by mistake or
a database directory is mounted to the NFS that sometimes has a
connection issue, there would be failures from the underlying
database layer. In this case, we should reopen database to get
a new handle.

Change-Id: Ib21de9dd7cc9c8b69abe82dcb2ed2a7011a96ca9
Reviewed-on: http://review.couchbase.org/16508
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>