MB-5458 Disable memory tracker if memory tracker stat is underflow
We observed that the memory accounting stats from the new memory
tracker suffer from underflow issues when a bucket is destroyed.
This would cause a node crash and rebalance failures in case of
multiple buckets. To avoid this issue, this change allows us to
switch to the old memory accounting stats for a bucket being
destroyed rather than borting the process.

As part of the future enhancement on memory accounting, we plan
to place memory usage barrier correctly across multiple
components.

Change-Id: Ic8b345af0891fff5f716a2ca5dfcf7b666f5d89f
Reviewed-on: http://review.couchbase.org/17033
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>