MB-5458 Disable memory tracker if memory tracker stat is underflow
We observed that the memory accounting stats from the new memory
tracker suffer from underflow issues when a bucket is destroyed.
This would cause a node crash and rebalance failures in case of
multiple buckets. To avoid this issue, this change allows us to
switch to the old memory accounting stats for a bucket being
destroyed rather than borting the process.

As part of the future enhancement on memory accounting, we plan
to place memory usage barrier correctly across multiple
components.

Change-Id: I8e48bbaa9e67faa9629e91e03d8243e1b4f40742
Reviewed-on: http://review.couchbase.org/17135
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>