CBD-377: Make synchronous vbucket delete optional
This change allows us to specify whether we want a synchronous
or asynchronous delete. When doing a synchronous delete we also
now put the connection into pending state to unblock the memcached
worker thread.

By default we do asynchronous vbucket deletes. This commit adds a
parameter in the value field that cansignal whether or not we
should do a vbucket delete asynchronously or synchronously. To
do a synchronous delete the value field should be set to "async=0"

Change-Id: If8ea49873305fc0fcade0581d8a4e7539a510c17
Reviewed-on: http://review.couchbase.org/17369
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>