MB-4870: t/binary-sasl fails when using isasl
We currently have a set of different sasl implementations being
used, and the test assumes to be able to run saslpasswd etc in
order to add auth for users. iSASL got it's own sasl database etc
so we should fix the test case. I believe it will be easier to
rewrite a new test in C where we got full control of the compile
flags memcached was built with

For now disable the t/binary-sasl test for iSASL

Change-Id: I67a2ad9e7b7bfbadead662cce7a5d092462adaf6
Reviewed-on: http://review.couchbase.org/17389
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>