MB-5164 Optimize background fetch by batching multiple gets
As background fetches get backed-up, there is an opportunity to
batch the reads from a single vbucket into one bulk read request.
This patch includes restructuring of the background fetch into a
batch of multiple gets. It also ensure that reads are scheduled
fairly by a new task, bgfetcher, for the read-only dispatcher.

Change-Id: I4e7a5f6122bfa7303a45c971508fae3266bc6067
Reviewed-on: http://review.couchbase.org/17593
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Jin Lim <jin@couchbase.com>