MB-4879 Use the previous session stats to determine TAP backfill
When a given node is restarted due to crash or shutdown,
all the TAP connections on that node will be established again.
If a replication from that node to one of the other nodes was not
completed before crash or shutdown, TAP backfill should be
scheduled to make both active and replica vbuckets consistent.

To address the above issue, this change logs TAP stats as part of
stats snapshot task and reads those TAP stats during warmup in
order to determine if backfill is required for a specific TAP
conneciton after warmup.

Change-Id: I638f67e09a25a7b0c009e0cae6d44058a88358d3
Reviewed-on: http://review.couchbase.org/17687
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>