MB-5718: Fixed race in flusher
If you try to stop and restart the flusher really quickly then
you can potentially run into an issue if the flusher is in the
pausing state (not paused) because going from pausing to running
is not considered a valid transition. This change fixes that
issue by making this a valid transition.

Change-Id: I69090053720b51df18925773405bb501ffa4fd2a
Reviewed-on: http://review.couchbase.org/17734
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>