MB-5746: Fix set with meta test case
When you set the expiration time of an item and the exptime is below
30 days then we set the expiration to that time relative to the current
time. Anything over 30 days is considered absolute time after the epoch.
Item meta data created in a test case doesn't do through the conversion for
values under 30 days. As a result the time here was being set to 5 minutes
after the epoch.

This is a problem for test cases because the item will expire if it is
accessed after being set and will throw off tests that aren't expecting
the item to expire.

In one of the tests there were two set with meta operations. If the
sequence was set-expire-set then the cas value would be one higher
than if the sequence was set-set-(expire on next call).

This change also adds logic to convert a relative expiration time
to its absolute time in the setWithMeta and delWithMeta funcitons.

Change-Id: I24550ac8a866d3bf6cf7549335d92e0e06bf36e7
Reviewed-on: http://review.couchbase.org/17784
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>