MB-5774: Check backfill completion when we are in paused state
This commit fixes an issue in the "tap stream" test case that caused
that test case to intermitently time out. The reason for the timeout
is a race condition that can occur if all items from backfill have
been removed from the queues (send to the consumer), but the
backfill stream hasn't yet notified the tap producer that it is
finished. If backfill completion comes in after the last backfill item
is sent then there is no code path that will mark the backfill as
completed and the tap stream will remain in a state of paused
forever.

This change add a call to the function that marks the backfill as
completed (if finished) in the code path for the tap paused state.

Change-Id: I4e67259d52b84ffd54290a2798c2506249a0d974
Reviewed-on: http://review.couchbase.org/17867
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>