MB-5790 Move the existing dirty item to the end of open checkpoint
Previously, we performed the deduplication by removing the
existing item and appending the new item to the end of the open
checkpoint. This might cause some more memory overhead because
the old existing item might be still referenced by the flusher or
the TAP producer.

This change simply moves the existing dirty item to the end of open
checkpoint for deduplication.

Change-Id: I20afb756697afa1473d46ac058c8457ca46889b8
Reviewed-on: http://review.couchbase.org/17894
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>