MB-5800 Don't create mc-engine instances for read-only kvstore
If CouchKVStore instance is read-only, we then don't need to
instantiate the mc-engine instance that allows us to synchronize
the db updates with the DB compactor and indexer running in
another process.

We plan to get rid of this mc-engine dependency from CouchKVStore
because this tight coupling between two components would
cause database inconsistency and make the extensibility quite
difficult.

Change-Id: I513c37b159b192c4e854873b02021b1a25f2639d
Reviewed-on: http://review.couchbase.org/17921
Reviewed-by: Liang Guo <liang@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>