MB-5624 Use TAP producer's creation time as its connection token
As a memcached's connection instance is reused for future
incomming TAP connection requests, it should not be used for the
TAP producer's connection token that is used to determine if a
bg fetched item is still valid and should be pushed into the queue
or not.

This change uses the TAP producer's creation time as its
connection token instead.

Change-Id: I58d0960968c6a67dfea425d03759c28901baf0b9
Reviewed-on: http://review.couchbase.org/18802
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>