[Backport] MB-5624 Use TAP producer's creation time as its token
As a memcached's connection instance is reused for future
incomming TAP connection requests, it should not be used for the
TAP producer's connection token that is used to determine if a
bg fetched item is still valid and should be pushed into the queue
or not.

This change uses the TAP producer's creation time as its
connection token instead.

Change-Id: I7e6a1164e15e14aea2eb180a0771cc52319daaaf
Reviewed-on: http://review.couchbase.org/19311
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>