MB-5935: Don't decr non-resident deleted items twice
There are many reasons that we can delete an item multiple
times. A couple are listed below:

1. And item can be set with a short expiration time. This can
cause a delete to be replicated due to expiration and a delete
to be replicated when the item actually is expired in that node.

2. XDCR can cause multiple deletes. An item can be deleted
locally and then just after this happens be deleted from the
remote cluster.

The problem is that when we delete an item we reset the value and
mark the item as non-resident since no value exists for the item.
If an item is deleted twice the non-resident item counter can be
decremented multiple times. This change makes sure an item is not
already deleted before decrementing the counter.

Change-Id: I0b24f803543825565bd8446eb717b8beb1989cca
Reviewed-on: http://review.couchbase.org/19042
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>