MB-6069: Remove sqlite tests from ep-engine
We will be deprecating the sqlite backend from ep-engine
in the near future. At the moment we have a few sporadic
sqlite testcase failures so we are going to remove them
from the testsuite. This change removes all sqlite only
tests and configures the remaining tests to be run only
against couchstore. Below is a list of tests that are being
permenately removed.

1. test init failure
2. test vbucket deletion doesn't affect new data
3. test restart with non-empty DB and empty mutation log
4. test db shards
5. test single db strategy
6. test single in-memory db strategy
7. restore: not enabled
8. restore: no such file
9. restore: invalid file
10. restore: data miss during restore
11. restore: with keys
12. restore: multiple incremental files

Change-Id: If2be1cc79b96a940347813c6b231606d6b556179
Reviewed-on: http://review.couchbase.org/19102
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>